# apps/api/openapi/schema.yml
openapi: 3.0.3
info:
  title: RankWant API
  version: 1.0.0
  description: Sport dasturlash va olimpiada platformasi — ochiq REST API
paths:
...
  /api/v1/attempts/:
    get:
      operationId: attempts_list
      description: Submit va urinishlar tarixi — PRD P0-3.
      parameters:
      - name: cursor
        required: false
        in: query
        description: The pagination cursor value.
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - attempts
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAttemptList'
          description: ''
    post:
      operationId: attempts_create
      description: Submit va urinishlar tarixi — PRD P0-3.
      tags:
      - attempts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttemptCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AttemptCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttemptCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
          description: ''
...
components:
  schemas:
...
    Attempt:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
        problem:
          type: string
          readOnly: true
        language:
          type: string
          readOnly: true
        verdict:
          $ref: '#/components/schemas/VerdictEnum'
        score:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        time_ms:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        memory_kb:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        failed_test_index:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        judged_at:
          type: string
          format: date-time
          nullable: true
      required:
      - created_at
      - id
      - language
      - problem
      - username
    AttemptCreate:
      type: object
      properties:
        problem:
          type: string
          pattern: ^[-a-zA-Z0-9_]+$
        language:
          type: string
          pattern: ^[-a-zA-Z0-9_]+$
        source_code:
          type: string
        contest:
          type: string
          nullable: true
          pattern: ^[-a-zA-Z0-9_]+$
      required:
      - language
      - problem
      - source_code
...
    PaginatedAttemptList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/Attempt'
...
    VerdictEnum:
      enum:
      - PENDING
      - RUNNING
      - AC
      - WA
      - TLE
      - MLE
      - OLE
      - RE
      - CE
      - PE
      - PARTIAL
      - IE
      - SKIPPED
      - COMPILE_TIMEOUT
      - IDLENESS
      - SECURITY_VIOLATION
      - CHECKER_ERROR
      - TESTING_ABORTED
      - RATE_LIMITED
      - DENIAL_OF_JUDGEMENT
      type: string
      description: |-
        * `PENDING` - Navbatda
        * `RUNNING` - Tekshirilmoqda
        * `AC` - Accepted
        * `WA` - Wrong Answer
        * `TLE` - Time Limit Exceeded
        * `MLE` - Memory Limit Exceeded
        * `OLE` - Output Limit Exceeded
        * `RE` - Runtime Error
        * `CE` - Compilation Error
        * `PE` - Presentation Error
        * `PARTIAL` - Qisman ball
        * `IE` - Internal Error
        * `SKIPPED` - O'tkazib yuborildi
        * `COMPILE_TIMEOUT` - Kompilyatsiya cho'zildi
        * `IDLENESS` - Idleness Limit Exceeded
        * `SECURITY_VIOLATION` - Xavfsizlik qoidasi buzildi
        * `CHECKER_ERROR` - Checker xatosi
        * `TESTING_ABORTED` - Tekshiruv bekor qilindi
        * `RATE_LIMITED` - Submit limiti
        * `DENIAL_OF_JUDGEMENT` - Infra nosozligi
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid